refactor: cleanup comparison pages
