--- src/app/features/country/components/country-comparison/country-comparison.component.html
@@ -1,4 +1,0 @@
-<app-country-menu [selected]="country | async" (changed)="switchCountry($event)"></app-country-menu>
-<app-country-menu [selected]="countryTwo | async" (changed)="switchCountryTwo($event)"></app-country-menu>
-<app-country-data [country]="country | async"></app-country-data>
-<app-country-data [country]="countryTwo | async"></app-country-data>
