+++ src/app/features/country/index.ts
@@ -1,2 +1,1 @@
-export * from './country-page.component';
-export * from './components/menu/menu.component';
+export * from './pages';
