+++ src/app/features/country/pages/comparison/comparison-page.component.scss
@@ -1,0 +1,4 @@
+.split {
+  display: flex;
+  justify-content: space-around;
+}
