+++ src/app/features/country/pages/index.ts
@@ -1,0 +1,2 @@
+export * from './comparison/comparison-page.component';
+export * from './country/country-page.component';
